backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
zope.interface==5.4.0
